Move kcl integration specs + TestKCLServiceSpec to receiver modules
